+++ app/src/main/java/com/example/cameraapp2/MainActivityUtilities.kt
@@ -1,22 +1,34 @@
+@file:JvmName("MainActivityUtilities")
+
 package com.example.cameraapp2
 
 import android.content.Context
+import android.graphics.Bitmap
+import android.view.View
 import androidx.appcompat.app.AlertDialog
-import androidx.camera.core.ImageCapture
-import androidx.camera.core.ImageCapture.OnImageCapturedCallback
-import androidx.camera.core.ImageProxy
-import java.util.concurrent.Executor
+import java.io.ByteArrayOutputStream
 
-fun ImageCapture.takePicture(
-    executor: Executor,
-    callback: OnImageCapturedCallback.(ImageProxy) -> Unit
-) {
-    this.takePicture(executor, object : OnImageCapturedCallback() {
-        override fun onCaptureSuccess(image: ImageProxy) {
-            super.onCaptureSuccess(image)
-            callback(image)
-        }
-    })
+fun cropImage(bitmap: Bitmap, frame: View?, reference: View?): ByteArray {
+    val heightOriginal = frame!!.height
+    val widthOriginal = frame.width
+    val heightFrame = reference!!.height
+    val widthFrame = reference.width
+    val leftFrame = reference.left
+    val topFrame = reference.top
+    val heightReal = bitmap.height
+    val widthReal = bitmap.width
+    val widthFinal = widthFrame * widthReal / widthOriginal
+    val heightFinal = heightFrame * heightReal / heightOriginal
+    val leftFinal = leftFrame * widthReal / widthOriginal
+    val topFinal = topFrame * heightReal / heightOriginal
+    val bitmapFinal = Bitmap.createBitmap(
+        bitmap, leftFinal, topFinal, widthFinal, heightFinal
+    )
+    val stream = ByteArrayOutputStream()
+    bitmapFinal.compress(
+        Bitmap.CompressFormat.JPEG, 100, stream
+    ) //100 is the best quality possible
+    return stream.toByteArray()
 }
 
 fun showPermissionsDeniedDialog(context: Context) {
