+++ app/src/main/java/com/example/cameraapp2/permissions/ScanImagePermissions.kt
@@ -1,3 +1,5 @@
+@file:JvmName("AppPermissions")
+
 package com.example.cameraapp2.permissions
 
 import android.Manifest
