+++ app/src/main/java/com/example/cameraapp2/textrecognition/Recognizer.java
@@ -1,4 +1,4 @@
-package com.example.cameraapp2;
+package com.example.cameraapp2.textrecognition;
 
 import android.graphics.Bitmap;
