+++ app/src/main/java/com/example/cameraapp2/textrecognition/RecognizerCallback.java
@@ -1,4 +1,4 @@
-package com.example.cameraapp2;
+package com.example.cameraapp2.textrecognition;
 
 public interface RecognizerCallback {
     void onCallBack(String stopName);
